refactor: 프로젝트 전반에 적용되는 css 파일들을 global 디렉터리로 이동
